Move enchants into separate package.
