+++ src/main/java/info/coremodding/craftenchants/enchants/item/enchants/ArmorKit.java
@@ -3,17 +3,18 @@
  * http://www.gnu.org/licenses/gpl.html
  * http://www.gnu.org/licenses/lgpl.html
  */
-package info.coremodding.craftenchants.enchants.item;
+package info.coremodding.craftenchants.enchants.item.enchants;
 
 import info.coremodding.craftenchants.enchants.ItemCE;
 
 import net.minecraft.enchantment.Enchantment;
 
-public class ReinforcedNock extends ItemCE {
-    public ReinforcedNock(int id) {
+public class ArmorKit extends ItemCE {
+
+    public ArmorKit(int id) {
         super(id);
-        setName("reinforcednock");
-        setEnchant(Enchantment.power, 2);
+        setName("armorkit");
+        setEnchant(Enchantment.unbreaking, 1);
     }
 
 }
