+++ src/main/java/info/coremodding/craftenchants/enchants/item/enchants/Counterweight.java
@@ -3,28 +3,29 @@
   * http://www.gnu.org/licenses/gpl.html
   * http://www.gnu.org/licenses/lgpl.html
   */
-package info.coremodding.craftenchants.enchants.item;
+package info.coremodding.craftenchants.enchants.item.enchants;
 
 import info.coremodding.craftenchants.enchants.ItemCE;
 
 import net.minecraft.enchantment.Enchantment;
 import net.minecraft.item.EnumToolMaterial;
 
-public class Studs extends ItemCE {
+public class Counterweight extends ItemCE {
     private EnumToolMaterial enchantMaterial;
-    
     /**
-     * @param id
-     * @param enchantMaterial
+     * Main constructor for Counterweight. Applied to a weapon, attributes Knockback.
+     * @param id Unique identifier for the item in the registry
+     * @param enchantMaterial Key ingredient the enchantment uses.
      */
-    public Studs(int id, EnumToolMaterial enchantMaterial) {
+    public Counterweight(int id, EnumToolMaterial enchantMaterial) {
         super(id);
         this.enchantMaterial = enchantMaterial;
-        setName(this.enchantMaterial.toString().toLowerCase() + "studs");
-        setEnchant(Enchantment.thorns, getEnchantLevel());
+        setName(this.enchantMaterial.toString().toLowerCase() + "counterweight");
+        setEnchant(Enchantment.knockback, getEnchantLevel());
     }
     
     public int getEnchantLevel() {
         return this.enchantMaterial.getHarvestLevel();
     }
+
 }
