+++ src/main/java/info/coremodding/craftenchants/enchants/item/enchants/ReinforcedNock.java
@@ -3,17 +3,17 @@
  * http://www.gnu.org/licenses/gpl.html
  * http://www.gnu.org/licenses/lgpl.html
  */
-package info.coremodding.craftenchants.enchants.item;
+package info.coremodding.craftenchants.enchants.item.enchants;
 
 import info.coremodding.craftenchants.enchants.ItemCE;
 
 import net.minecraft.enchantment.Enchantment;
 
-public class SharpeningStone extends ItemCE {
-    
-    public SharpeningStone(int id) {
+public class ReinforcedNock extends ItemCE {
+    public ReinforcedNock(int id) {
         super(id);
-        setName("sharpeningstone");
-        setEnchant(Enchantment.sharpness, 1);
+        setName("reinforcednock");
+        setEnchant(Enchantment.power, 2);
     }
+
 }
