+++ src/main/java/info/coremodding/craftenchants/enchants/item/enchants/SharpeningStone.java
@@ -3,17 +3,17 @@
  * http://www.gnu.org/licenses/gpl.html
  * http://www.gnu.org/licenses/lgpl.html
  */
-package info.coremodding.craftenchants.enchants.item;
+package info.coremodding.craftenchants.enchants.item.enchants;
 
 import info.coremodding.craftenchants.enchants.ItemCE;
 
 import net.minecraft.enchantment.Enchantment;
 
-public class Snorkle extends ItemCE {
-    public Snorkle(int id) {
+public class SharpeningStone extends ItemCE {
+    
+    public SharpeningStone(int id) {
         super(id);
-        setName("snorkle");
-        setEnchant(Enchantment.respiration, 2);
+        setName("sharpeningstone");
+        setEnchant(Enchantment.sharpness, 1);
     }
-
 }
