+++ src/main/java/info/coremodding/craftenchants/enchants/item/enchants/Studs.java
@@ -3,29 +3,28 @@
   * http://www.gnu.org/licenses/gpl.html
   * http://www.gnu.org/licenses/lgpl.html
   */
-package info.coremodding.craftenchants.enchants.item;
+package info.coremodding.craftenchants.enchants.item.enchants;
 
 import info.coremodding.craftenchants.enchants.ItemCE;
 
 import net.minecraft.enchantment.Enchantment;
 import net.minecraft.item.EnumToolMaterial;
 
-public class Counterweight extends ItemCE {
+public class Studs extends ItemCE {
     private EnumToolMaterial enchantMaterial;
+    
     /**
-     * Main constructor for Counterweight. Applied to a weapon, attributes Knockback.
-     * @param id Unique identifier for the item in the registry
-     * @param enchantMaterial Key ingredient the enchantment uses.
+     * @param id
+     * @param enchantMaterial
      */
-    public Counterweight(int id, EnumToolMaterial enchantMaterial) {
+    public Studs(int id, EnumToolMaterial enchantMaterial) {
         super(id);
         this.enchantMaterial = enchantMaterial;
-        setName(this.enchantMaterial.toString().toLowerCase() + "counterweight");
-        setEnchant(Enchantment.knockback, getEnchantLevel());
+        setName(this.enchantMaterial.toString().toLowerCase() + "studs");
+        setEnchant(Enchantment.thorns, getEnchantLevel());
     }
     
     public int getEnchantLevel() {
         return this.enchantMaterial.getHarvestLevel();
     }
-
 }
